separate Prawn class from dispensary controller
